Rename StoredConsent to StoredConsentImage
The previous StoredConsent data model is only relevant
to the ConsentHistory DynamoDB data model's oldImage
and newImage attributes which represent the subset of
consent attributes stored in consent images in the
consent history data store.

Renamed from StoredConsent to StoredConsentImage and
moved to /infrastructure/entities to better reflect
the data model scope and purpose.