Update ServiceUserConsent DDB models to reflect updated GSI
The prior ActiveConsentsWithExpiryTime GSI had a design flaw
that it relied on being able to scan the DynamoDB GSI in
sorted order of the GSI sort key, expiryTime, in order to
efficiently determine which consents to expire, with the
GSI partition key autoExpireId having the consent partition
key if the consent is active and has an expiry time.

However, DynamoDB Scans do not support ordering results, and
in order to identify all consents expiring before the current
time, we would need to scan the full GSI, which does not scale
as our data set grows in size.

Replacing with a ActiveConsentsByExpiryHour GSI which uses the
consent expiry hour as its GSI partition key, and an expiryTimeId
attribute that appends the expiryTime + "|" + id, which will
allow us to have a periodic job query each expiry hour bucket in
the last few days and have all query results automatically sorted
in ascending order of expiry time.

Ref: Consent-Management-Platform/consent-management-api-cdk/#8